+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.5.6
+
+- `PerfectShape::Path#intersect?(rectangle)`
+
 ## 0.5.5
 
 - `PerfectShape::Arc#intersect?(rectangle)`